Change the GraphQL API files structure
